sysutils: Add neofetch
Neofetch is a command-line system information tool written in bash.

Neofetch displays information about your operating system, software and
hardware in an aesthetic and visually pleasing way.

The overall purpose of Neofetch is to be used in screen-shots of your system.
Neofetch shows the information other people want to see. There are other
tools available for proper system statistic/diagnostics.