py-wrapt: updated to 1.12.1
Version 1.12.1

Bugs Fixed

Applying a function wrapper to a static method of a class using the wrap_function_wrapper() function, or wrapper for the same, wasn't being done correctly when the static method was the immediate child of the target object. It was working when the name path had multiple name components. A failure would subsequently occur when the static method was called via an instance of the class, rather than the class.


Version 1.12.0

Features Changed

Provided that you only want to support Python 3.7, when deriving from a base class which has a decorator applied to it, you no longer need to access the true type of the base class using __wrapped__ in the inherited class list of the derived class.
Bugs Fixed

When using the synchronized decorator on instance methods of a class, if the class declared special methods to override the result for when the class instance was tested as a boolean so that it returned False all the time, the synchronized method would fail when called.
When using an adapter function to change the signature of the decorated function, inspect.signature() was returning the wrong signature when an instance method was inspected by accessing the method via the class type.